Send motor commands with both custom flight mode and Copter standard flight modes.
Motor commands can now be sent by both standard Copter flight modes
and the new custom flight mode. Therefore, the custom flight mode
must be detected in the Copter main loop. This is achieved by name4
identification (for the custom flight mode, name4 is 'XXXX').
Depending on the detected flight mode different functions are called.
For all flight modes the motors_output() function is called to send
motor commands.

However, the motors_output() function also had to be modified.
If the custom flight mode is active, the new function output_custom()
is called instead of output().
The output_custom() function is a copy of the output() function with
two modifications: 1. The output_boost_throttle() function must not be called.
2. Instead of the output_to_motors() function, the new function
output_to_motors_custom() is called.
The function output_to_motors_custom() is a copy of output_to_motors(), however,
in the custom function the new variable _custom_input is assigned to the actuator.
The _custom_input is set by the custom flight mode.