Fix controller to PWM output interface.
This version was tested in SITL and works as expected.
However, the motors are spinning although the vehicle is disarmed.
This must be changed in future commits.

There were multiple problems that were fixed:
One problem was that the wrong function was used to send the controller
outputs to the SRV_Channels. Moreover, the PWM of the servos was manipulated
in Attitude.cpp which was not intended. Additionally, the PWM of the motors
was manipulated by servos.cpp which was not intended either.
For a more general use, the controller outputs are now connected with
the index of a channel instead of a function (which was hard coded).
Therefore, the controller output must be connected with the correct channel.
The function used to send the controller outputs to the SRV_Channels uses
a normalized control signal. The normalized control signal is 0~1 for motors
and -1~1 for servos. Due to this difference, a function was written in SRV_Channel
to get the type of the actuator.