Remove all '-Wfloat-equal' from boards.py
The code did not compile for hardware targets if MatlabController
includes control allocation.
It is because the backslash operator of Matlab produces bad code...
As long as we have no better solution, we have to live with this.
But we should implement an alternative algortihm to solve linear
systems and then enable again the '-Wfloat-equal'.