Revert "changes to SIM_gazebo"
This reverts commit 6240795fa03e831d314d7104cdf88613b882b0f4.
The commit seemed to work in the past but it now produces an error.
After 2 seconds the SITL crashes with:

Connection reset or closed by peer on TCP socket
Attempting reconnect
[Errno 111] Connection refused sleeping

This error disappears by reverting the commit.
Moreover, the hard coded sample time of "0.025 seconds" seems to
be wrong since we usually use 0.0025 seconds.