Make the logging interface more convenient
The logging interface Simulink block "log muxer" only had
a capacity of 15 signals. Moreover, the logged signals
were labeled with generic names (v1,v2,v3,...) so that
the revision of logs was annoying.
The problem was solved by making the logging interface
block "log muxer" more complex. It is now possible to
change the number of inport, the logging labels and the
log batch labels for up to 5 batches of each 14 signals.