Improve EKF performance in case of rotor failures
By comparing the states of the EKF with the simulated
states it was found that there are relatively large
errors in the attitude and acceleration. Moreover,
after enabling the deactivated rotor again, it took
much time to reduce the attitude error (because the
EKF feedback gain was still small).
The problem was solved by creating a continuous
reduction of the EKF feedback gain depending on the
attitude error. The reduction only is active in case
of a rotational rate greater than 4 rad/s. With this
modification the attitude error can quickly be reduced
when the vehicle is not spinning.