Fix performance logging and change debug interface
Performance logging did not work because the
time variables had the wrong data type.

The debug interface had be changed to also
debug the control commands. However, the
debug interface had not been adjusted on the
LADAC side.
Since adjustments on the LADAC side have to
be done manually, it is probably better to
only include the measure struct. For the
cmd struct a separate debugging interface
could be added in the future.