Fix errors by merging so that ArduCopter compiles again
In Copter.cpp, a merge conflict was not resolved in
the previous commit because the basic structure of
the scheduler changed, making it difficult to update.
In this commit, the ArduCopter default scheduler is
restored so that the code can be compiled again.
However, the custom flight mode cannot be used.
his must be fixed in subsequent commits.
In addition, there was an error in AP_Motors_Class.h
that occurred during automatic merging: the function
get_type_string was duplicated.