Restore the custom flight mode
Since the ArduCopter scheduler is called differently
in Copter-4.4 than in Copter-4.2.0, it was not so easy
to add the custom flight mode.

The scheduler always calls the rate controller
(run_rate_controller), no matter which flight mode is
active, and sends signals to the motors (motors_output)
directly afterwards. This was also the case in
Copter-4.2.0 and is not desired in case the custom flight
mode is active. Previously it was prevented that
run_rate_controller and motors_output were called. This
is no longer so easy with the new scheduler. Therefore
the following solution was implemented:
The ArduCopter scheduler was not changed, so
run_rate_controller and motors_output are also called
in custom flight mode. Inside run_rate_controller and
motors_output the code was changed so that the rest of
the code inside the functions is not called when custom
flight mode is active. Sending the motor signals in
custom flight mode is now done directly in mode_custom.

With this commit,  the custom flight mode (mode 29) can
now be reactivated. In the SITL it was briefly tested
that the loiter flight mode and the guided flight mode
still work with LindiCopter. But it should be tested
more detailed.
But there was a problem with the arming in the SITL
(with --mode=gazebo):
AP: Arm: Main loop slow (xxxHz < 400Hz)
The problem only occurs sporadically when the frequency
of the UDP packets sent to the ArduCopter SITL is
increased to 1200Hz. Nevertheless, this problem should
be investigated in the future.