Fix too many MatlabControllerClass instances
'custom_controller' (instance of class 'MatlabControllerClass')
was member of class 'Mode'. This is not desired, because the
class 'Mode' is the base class for all derived flight mode classes.
In this way each flight mode instance also contained an instance of
'MatlabControllerClass' which means that there were as many
instances of 'MatlabControllerClass' as flight modes in ArduPilot.

'custom_controller' is now a private member of the derived class
'ModeCustom' and also the optional debugging socket has been placed here.

See also: 911cfe2c