Avoid FC reboot due to buggy UART interface
There were FC reboots when the distributed IMU boards
were connected. The reboots especially occurred when
the mode_custom.cpp took much computation time. In
this case, the UART interface does not really work: The
IMU board LEDs flicker, which means that they do not
receive the trigger messages from the FC.
The reboots are avoided by this commit by removing the
"flush" functions. Moreover, the UART read decoding was
overworked to avoid potential FC reboots in case of
corrupted UART messages.
The UART interface runs more stable if the computation
time of the mode_custom.cpp is reduced. Moreover, it
helps to use the ArduPilot parameter SERIALx_PROTOCOL
27 because it seems to have a higher priority. However,
it only seems to work for one UART.