Resolve obvious merge conflicts (take 'Copter-4.6.0')
There are a lot of files with conflicts although the
custom flight mode only changed a few files.
That is why, most of the conflicts of most files, i.e.
all the files that were not changed for the custom
flight mode, could be easily solved just by taking
the files from 'Copter-4.6.0'.
The remaining conflicts were not solved yet. These files
are:
- AP_Mission.cpp
- AP_Mission.h
- AP_Motors_Class.h
- AP_NavEKF3_core.cpp
- Attitude.cpp
- boards.py
- Copter.cpp
- Copter.h
- MissionItemProtocol_Waypoints.cpp
- mode.h
- motors.cpp
- Parameters.cpp

These conflicts must be solved in the next commit.