Removal of the generated file ModeCustomParams.cpp
This file was needed for tunable parameters of
the Matlab controller. It caused a compiler error
because it included Copter.h which did not work
except ModeCustomParams.cpp was moved to the
ArduCopter folder. The file only contains the
definition of ModeCustom::var_info[], which was
moved to mode_custom.cpp. mode_custom.cpp is
located in the ArduCopter folder and already
includes Copter.h.